+++ cmake/modules/CheckCXXFeatures.cmake
@@ -10,15 +10,6 @@
 #    This module internally sets the following variables, which are then
 #    exported into the config.h of the current dune module.
 #
-#    :code:`HAS_ATTRIBUTE_UNUSED`
-#       True if attribute unused is supported
-#
-#    :code:`HAS_ATTRIBUTE_DEPRECATED`
-#       True if attribute deprecated is supported
-#
-#    :code:`HAS_ATTRIBUTE_DEPRECATED_MSG`
-#       True if attribute deprecated("msg") is supported
-#
 # .. cmake_variable:: DISABLE_CXX_VERSION_CHECK
 #
 #    You may set this variable to TRUE to disable checking for
@@ -185,82 +176,6 @@ endif()
 
 # make sure we have at least C++17
 dune_require_cxx_standard(MODULE "DUNE" VERSION 17)
-
-# perform tests
-
-# __attribute__((unused))
-check_cxx_source_compiles("
-   int main(void)
-   {
-     int __attribute__((unused)) foo;
-     return 0;
-   };
-"  HAS_ATTRIBUTE_UNUSED
-)
-
-# __attribute__((deprecated))
-check_cxx_source_compiles("
-#define DEP __attribute__((deprecated))
-   class bar
-   {
-     bar() DEP;
-   };
-
-   class peng { } DEP;
-
-   template <class T>
-   class t_bar
-   {
-     t_bar() DEP;
-   };
-
-   template <class T>
-   class t_peng {
-     t_peng() {};
-   } DEP;
-
-   void foo() DEP;
-
-   void foo() {}
-
-   int main(void)
-   {
-     return 0;
-   };
-"  HAS_ATTRIBUTE_DEPRECATED
-)
-
-# __attribute__((deprecated("msg")))
-check_cxx_source_compiles("
-#define DEP __attribute__((deprecated(\"message\")))
-   class bar {
-     bar() DEP;
-   };
-
-   class peng { } DEP;
-
-   template <class T>
-   class t_bar
-   {
-     t_bar() DEP;
-   };
-
-   template <class T>
-   class t_peng
-   {
-     t_peng() {};
-   } DEP;
-
-   void foo() DEP;
-
-   void foo() {}
-
-   int main(void)
-   {
-     return 0;
-   };
-"  HAS_ATTRIBUTE_DEPRECATED_MSG
-)
 
 # ******************************************************************************
 #